backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
motor>=3.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
cryptography>=41.0.0
cloudinary>=1.36.0
python-multipart>=0.0.9
slowapi>=0.1.9
python-magic>=0.4.27
apscheduler>=3.10.0
sendgrid>=6.11.0
firebase-admin>=6.2.0
googlemaps>=4.10.0
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
